+++ obj/src/clients/AzureFunctionClient.js
@@ -33,6 +33,10 @@ const AzureConnectionResolver_1 = require("../connect/AzureConnectionResolver");
  *     - protocol:                    (optional) connection protocol
  *     - app_name:                    (optional) Azure Function application name
  *     - function_name:               (optional) Azure Function name
+ * - options:
+ *      - retries:               number of retries (default: 3)
+ *      - connect_timeout:       connection timeout in milliseconds (default: 10 sec)
+ *      - timeout:               invocation timeout in milliseconds (default: 10 sec)
  * - credentials:
  *     - auth_code:                   Azure Function auth code if use custom authorization provide empty string
  *
@@ -74,7 +78,7 @@ const AzureConnectionResolver_1 = require("../connect/AzureConnectionResolver");
  */
 class AzureFunctionClient {
     constructor() {
-        this._retries = 1;
+        this._retries = 3;
         /**
          * The default headers to be added to every request.
          */
@@ -117,6 +121,8 @@ class AzureFunctionClient {
         this._connectionResolver.configure(config);
         this._dependencyResolver.configure(config);
         this._connectTimeout = config.getAsIntegerWithDefault('options.connect_timeout', this._connectTimeout);
+        this._retries = config.getAsIntegerWithDefault("options.retries", this._retries);
+        this._timeout = config.getAsIntegerWithDefault("options.timeout", this._timeout);
     }
     /**
      * Sets references to dependent components.