+++ obj/src/clients/CommandableAzureFunctionClient.d.ts
@@ -13,6 +13,10 @@ import { AzureFunctionClient } from './AzureFunctionClient';
  *     - protocol:                    (optional) connection protocol
  *     - app_name:                    (optional) Azure Function application name
  *     - function_name:               (optional) Azure Function name
+ * - options:
+ *      - retries:               number of retries (default: 3)
+ *      - connect_timeout:       connection timeout in milliseconds (default: 10 sec)
+ *      - timeout:               invocation timeout in milliseconds (default: 10 sec)
  * - credentials:
  *     - auth_code:                   Azure Function auth code if use custom authorization provide empty string
  *