+++ obj/src/clients/CommandableAzureFunctionClient.js
@@ -25,6 +25,10 @@ const AzureFunctionClient_1 = require("./AzureFunctionClient");
  *     - protocol:                    (optional) connection protocol
  *     - app_name:                    (optional) Azure Function application name
  *     - function_name:               (optional) Azure Function name
+ * - options:
+ *      - retries:               number of retries (default: 3)
+ *      - connect_timeout:       connection timeout in milliseconds (default: 10 sec)
+ *      - timeout:               invocation timeout in milliseconds (default: 10 sec)
  * - credentials:
  *     - auth_code:                   Azure Function auth code if use custom authorization provide empty string
  *
@@ -85,14 +89,12 @@ class CommandableAzureFunctionClient extends AzureFunctionClient_1.AzureFunction
             const timing = this.instrument(correlationId, this._name + '.' + cmd);
             try {
                 const result = yield this.call(cmd, correlationId, params);
+                timing.endTiming();
                 return result;
             }
             catch (err) {
                 timing.endFailure(err);
                 return err;
-            }
-            finally {
-                timing.endTiming();
             }
         });
     }