+++ src/clients/AzureFunctionClient.ts
@@ -32,6 +32,10 @@ import { AzureConnectionResolver } from '../connect/AzureConnectionResolver';
  *     - protocol:                    (optional) connection protocol
  *     - app_name:                    (optional) Azure Function application name
  *     - function_name:               (optional) Azure Function name
+ * - options:
+ *      - retries:               number of retries (default: 3)
+ *      - connect_timeout:       connection timeout in milliseconds (default: 10 sec)
+ *      - timeout:               invocation timeout in milliseconds (default: 10 sec)
  * - credentials:
  *     - auth_code:                   Azure Function auth code if use custom authorization provide empty string
  *  
@@ -81,7 +85,7 @@ export abstract class AzureFunctionClient implements IOpenable, IConfigurable, I
      */
     protected _connection: AzureConnectionParams;
 
-    protected _retries: number = 1;
+    protected _retries: number = 3;
     /**
      * The default headers to be added to every request.
      */
@@ -130,6 +134,8 @@ export abstract class AzureFunctionClient implements IOpenable, IConfigurable, I
 		this._dependencyResolver.configure(config);
 
         this._connectTimeout = config.getAsIntegerWithDefault('options.connect_timeout', this._connectTimeout);
+        this._retries = config.getAsIntegerWithDefault("options.retries", this._retries);
+        this._timeout = config.getAsIntegerWithDefault("options.timeout", this._timeout);
     }
 
     /**