+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200">
 <br/> Azure specific components for Node.js / ES2017 Changelog
 
+## <a name="1.0.3"></a> 1.0.3 (2022-09-07)
+
+### Features
+* Added interceptors by command or regex
+* Added more configs for clients
+
 ## <a name="-1.0.1-1.0.2"></a> 1.0.1-1.0.2 (2022-03-09)
 
 ### Bug Fixes